// src/app/modules/Old_JobList/Components/Interfaces.tsx
 export interface DataRow {
    createdAt: number;
    updatedAt: number;
    id: number;
    task_type: number;
    cleanerid: number;
    attendencedate: Date;
    customerids: string;
    orderid: number;
    neworderid: number;
    attendencedatedone: string;
    attendence_server_datedone: string;
    markedby: number;
    latitude: string;
    longitude: string;
    address: string;
    status: number;
    servicetype: number;
    timeslot: string;
    point: number;
    notathome: number;
    job_status_id: number;
    markby_usertype: number;
    siteid: number;
    method: string;
    date_changed_from: string;
    name: string;
    s_latitude: string;
    s_longitude: string;
    phone: string;
    caddress: string;
    city: number;
    vehicleid: number;
    timeslotname: string;
    time_from: string;
    time_to: string;
    c_id: number;
    c_fname: string;
    c_lname: string;
    c_phone: string;
    c_onwork: number;
    customercity: string;
    ctJobRating: null;
    att_job_status: string;
    societyname: string;
    customerdata: Customerdata;
    servicetypename: string;
    vehiclename: string;
  }
  export interface Customerdata {
    id: number;
    cleaner_id: number;
    gst_number: number;
    distributorid: number;
    first_name: string;
    last_name: string;
    email: string;
    phone: string;
    gender: string;
    flatno: string;
    locality: string;
    address_line_1: string;
    address_line_2: string;
    latitude: string;
    longitude: string;
    city: number;
    state: string;
    pincode: string;
    is_premium: number;
    termscondi: number;
    status: number;
    profile_image: string;
    stage: number;
    societyid: number;
    cleanercode: string;
    cleanerlead: number;
    is_autodebit_enable: number;
    userType: number;
    auto_debit_access_key: null;
    customer_authentication_id: null;
    authorization_status: null;
    coins: number;
    wallet_coins: number;
    razorpay_customer_id: string;
    referBy: string;
    app_rating: number;
    in_serviceable_area: number;
    is_deleted: number;
    other_society_name: string;
    source_id: number;
    enable_notification: number;
    campaign_id: number;
    active_status: number;
    createdAt: number;
    updatedAt: number;
  }